Fixed bug with matching case-sensitive custom smileys. It happened because string compare does not support -exact.
It replaced it with static -nocase because in this check is only a workaround for a bug in tcl with elicit. We want to make sure that there is nothing embdeded on sides or something like that, not if the two things are exactly the same. That is done already.


git-svn-id: https://amsn.svn.sourceforge.net/svnroot/amsn/trunk@6281 f248a3e1-9e13-0410-b2cf-a4429de6d898