updated the feature: colored text in chatwindow. Now the feature is in WinWrite proc. It's more fast thanks to an "if control" to see if we really need to call substSmiley proc, because with a faded nick substSmiley proc was called to each letter...
Now WinWrite accepts a new (default) value: lst (a list).

Fixed an old bug which allowed to draw custom emoticons in the nickname.

In this commit there isn't identation, in this way is more simple to see 
the real difference. I will do another commit for identation, but not 
today.



git-svn-id: https://amsn.svn.sourceforge.net/svnroot/amsn/trunk@10422 f248a3e1-9e13-0410-b2cf-a4429de6d898