Refactored libtray to make it more compliant with specs
The creation of the Tk Window that will hold the tray icon is delayed
until a system tray is available. This required some refactoring to keep
the same functionality. In particular a -createcb flag has been added
to both newti and configureti because otherwise it would be impossible
to bind mouse events to the window - if the window does not yet exist...
Also some code has been updated/optimized/simplified.